requirements.txt
python-dotenv
pandas 
numpy 
certifi
pymongo
pymongo
scikit-learn
dill
pyaml
mlflow
dagshub
fastapi
uvicorn
pytorch
torchvision
##-e .